pyproject.toml
[project]
name = "devtoday-backend"
version = "0.1.0"
description = "DevelopersToday test task (backend part)"
authors = [
    {name = "TeraBasedProgrammer",email = "ilyadronov2332@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

